feat: bring the engine core onto capstone v6
Points the dependency at capstone v6 and adapts the architecture-independent
core to it. `cs_option()`'s value parameter became uintptr_t upstream (#2845),
which Swift imports as UInt while the previous `size_t` imports as Int; seven
call sites in Options.swift are converted, four of them via the shared
`Bool.csOptValue` helper.

`adaptedArchitectures` is empty for now. Every architecture wrapper is still
v5 code, so including any of them here would not compile. The default build is
therefore core-only and compilable; AArch64 is added back once its wrapper is
rewritten. Traits stay declared, so an architecture can still be enabled
explicitly — it just will not build yet.

This narrows the error surface from 41 sites across 8 files to 19 sites in
Arm64.swift alone.

Steps 2 and 3 of Documentations/Evolutions/draft-adapt-capstone-v6.md, merged
so that the commit builds. The proposal records why the steps were re-sliced,
and a capstone v5 SwiftPM defect found on the way (TriCore cannot find
platform.h there; v6 is unaffected).

+++ Sources/Capstone/Options.swift
@@ -77,7 +77,7 @@ extension Capstone {
         let err: cs_err
         switch option {
         case .syntax(syntax: let syntax):
-            err = cs_option(handle, CS_OPT_SYNTAX, syntax.rawValue)
+            err = cs_option(handle, CS_OPT_SYNTAX, uintptr_t(syntax.rawValue))
         case .detail(value: let value):
             err = cs_option(handle, CS_OPT_DETAIL, value.csOptValue)
             if err == CS_ERR_OK {
@@ -86,7 +86,7 @@ extension Capstone {
         case .unsigned(value: let value):
             err = cs_option(handle, CS_OPT_UNSIGNED, value.csOptValue)
         case .mode(mode: let mode):
-            err = cs_option(handle, CS_OPT_MODE, Int(mode.rawValue))
+            err = cs_option(handle, CS_OPT_MODE, uintptr_t(mode.rawValue))
         case .skipDataEnabled(let enabled):
             err = cs_option(handle, CS_OPT_SKIPDATA, enabled.csOptValue)
         case .skipData(mnemonic: let mnemonic, callback: let callback):
@@ -97,7 +97,7 @@ extension Capstone {
             }
             err = mnemonic.withCString { mnemonicPtr in
                 withUnsafePointer(to: cs_opt_mnem(id: instruction.rawValue, mnemonic: mnemonicPtr)) {
-                    cs_option(handle, CS_OPT_MNEMONIC, Int(bitPattern: $0))
+                    cs_option(handle, CS_OPT_MNEMONIC, uintptr_t(bitPattern: $0))
                 }
             }
         }
@@ -132,7 +132,7 @@ extension Capstone {
                                                      user_data: Unmanaged.passUnretained(self).toOpaque()
         ), {
             cs_option(handle, CS_OPT_SKIPDATA, true.csOptValue)
-            return cs_option(handle, CS_OPT_SKIPDATA_SETUP, Int(bitPattern: $0))
+            return cs_option(handle, CS_OPT_SKIPDATA_SETUP, uintptr_t(bitPattern: $0))
         })
     }
 
@@ -164,7 +164,8 @@ internal extension UnsafeMutablePointer where Pointee == Int8 {
 }
 
 fileprivate extension Bool {
-    var csOptValue: Int {
-        return Int((self ? CS_OPT_ON : CS_OPT_OFF).rawValue)
+    /// cs_option()'s value parameter is uintptr_t, which Swift imports as UInt.
+    var csOptValue: uintptr_t {
+        return uintptr_t((self ? CS_OPT_ON : CS_OPT_OFF).rawValue)
     }
 }
